Fixed bug

+++ pytracking/analysis/extract_results.py
@@ -159,7 +159,7 @@ def extract_results(trackers, dataset, report_name, skip_missing_seq=False, plot
             ave_success_rate_plot_center[seq_id, trk_id, :] = (err_center.view(-1, 1) <= threshold_set_center.view(1, -1)).sum(0).float() / seq_length
             ave_success_rate_plot_center_norm[seq_id, trk_id, :] = (err_center_normalized.view(-1, 1) <= threshold_set_center_norm.view(1, -1)).sum(0).float() / seq_length
 
-    print('\n\nComputed results over {} / {} sequences'.format(valid_sequence.sum().item(), valid_sequence.shape[0]))
+    print('\n\nComputed results over {} / {} sequences'.format(valid_sequence.long().sum().item(), valid_sequence.shape[0]))
 
     # Prepare dictionary for saving data
     seq_names = [s.name for s in dataset]
